deal with files with only a header line and no data without error (#75)
closes #73
supersedes #74 

This PR contains a lot to get to the root of the problem and not
duplicate code bits that are used in multiple places.

1. Three new, unexported, utility functions are added (with tests)
`has_header()`, `determine_header_types()`, and `read_first_line()`, and
they are utilized in other functions.
2. `guess_numerical_mark()` is made more robust so that it will not
error on a file with a single line (either header only or data only),
and tests are added to keep it that way. One class of problem that I
don't have a solution for is a CSV file with one line of data (no
header) that has a numerical mark (used in an unquoted numeric) that
collides with the delimiter, e.g. `XS1088274672,1,000.34,USD`. There is
ambiguity in there that I can't resolve.
3. `read_portfolio_csv()` is made more robust so that it will not error
on a file with only a header and no data (returns as expected NA) and
tests for that are added.
4. Also introduced the idea of optionally passing `encoding` and
`delimiter` to these utility functions so that they do not have to be
determined so many times during the process of reading a portfolio CSV.
Eventually, this concept should spread to other functions.